+++ residential/solvers.cpp
@@ -10,7 +10,7 @@
 #include "gridlabd.h"
 #include "solvers.h"
 
-static unsigned long _nan[2] = {0xffffffff,0x7fffffff};
+static uint32 _nan[2] = {0xffffffff,0x7fffffff};
 static double NaN = *(double*)&_nan;
 #define EVAL(t,a,n,b,m,c) (a*exp(n*t) + b*exp(m*t) + c)
 /** solve an equation of the form $f[ ae^{nt} + be^{mt} + c = 0 $f]
